Commit 4- Building the Pages
