+++ worldwise/src/App.jsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Product from "./pages/Product";
 import Pricing from "./pages/Pricing";
 import HomePage from "./pages/HomePage";
+import Login from "./pages/Login";
 import PageNotFound from "./pages/PageNotFound";
 import AppLayOut from "./pages/AppLayOut";
 
@@ -15,6 +16,7 @@ function App() {
           <Route path="/" element={<HomePage />} />
           <Route path="Product" element={<Product />} />
           <Route path="Pricing" element={<Pricing />} />
+          <Route path="/login" element={<Login />} />
           <Route path="app" element={<AppLayOut />} />
           <Route path="*" element={<PageNotFound />} />
         </Routes>
