+++ worldwise/src/components/PageNav.module.css
@@ -1,11 +1,34 @@
 .nav {
-  background-color: orangered;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 }
+
 .nav ul {
   list-style: none;
   display: flex;
-  justify-content: space-between;
+  align-items: center;
+  gap: 4rem;
 }
-.nav :global(.active) {
-  background-color: green;
+
+.nav a:link,
+.nav a:visited {
+  text-decoration: none;
+  color: var(--color-light--2);
+  text-transform: uppercase;
+  font-size: 1.5rem;
+  font-weight: 600;
+}
+
+/* CSS Modules feature */
+.nav a:global(.active) {
+  color: var(--color-brand--2);
+}
+
+a.ctaLink:link,
+a.ctaLink:visited {
+  background-color: var(--color-brand--2);
+  color: var(--color-dark--0);
+  padding: 0.8rem 2rem;
+  border-radius: 7px;
 }
