+++ worldwise/src/pages/HomePage.jsx
@@ -1,16 +1,26 @@
 import { Link } from "react-router-dom";
+import styles from "./Homepage.module.css";
 import PageNav from "../components/PageNav";
-import AppNav from "../components/AppNav";
 
-function HomePage() {
+export default function Homepage() {
   return (
-    <div>
+    <main className={styles.homepage}>
       <PageNav />
-      <AppNav />
-      <h1>WorldWise</h1>
-      <Link to="/app">go to the app</Link>
-    </div>
+      <section>
+        <h1>
+          You travel the world.
+          <br />
+          WorldWise keeps track of your adventures.
+        </h1>
+        <h2>
+          A world map that tracks your footsteps into every city you can think
+          of. Never forget your wonderful experiences, and show your friends how
+          you have wandered the world.
+        </h2>
+        <Link to="/app" className="cta">
+          Start tracking now
+        </Link>
+      </section>
+    </main>
   );
 }
-
-export default HomePage;
