+++ worldwise/src/pages/Login.jsx
@@ -1,4 +1,6 @@
+import { useState } from "react";
 import styles from "./Login.module.css";
+import PageNav from "../components/PageNav";
 
 export default function Login() {
   // PRE-FILL FOR DEV PURPOSES
@@ -7,6 +9,7 @@ export default function Login() {
 
   return (
     <main className={styles.login}>
+      <PageNav />
       <form className={styles.form}>
         <div className={styles.row}>
           <label htmlFor="email">Email address</label>
