--- worldwise/starter/components/AppNav.module.css
@@ -1,28 +1,0 @@
-.nav {
-  margin-top: 3rem;
-  margin-bottom: 2rem;
-}
-
-.nav ul {
-  list-style: none;
-  display: flex;
-  background-color: var(--color-dark--2);
-  border-radius: 7px;
-}
-
-.nav a:link,
-.nav a:visited {
-  display: block;
-  color: inherit;
-  text-decoration: none;
-  text-transform: uppercase;
-  font-size: 1.2rem;
-  font-weight: 700;
-  padding: 0.5rem 2rem;
-  border-radius: 5px;
-}
-
-/* CSS Modules feature */
-.nav a:global(.active) {
-  background-color: var(--color-dark--0);
-}
