--- worldwise/starter/components/Logo.jsx
@@ -1,7 +1,0 @@
-import styles from "./Logo.module.css";
-
-function Logo() {
-  return <img src="/logo.png" alt="WorldWise logo" className={styles.logo} />;
-}
-
-export default Logo;
