--- worldwise/starter/components/PageNav.module.css
@@ -1,34 +1,0 @@
-.nav {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.nav ul {
-  list-style: none;
-  display: flex;
-  align-items: center;
-  gap: 4rem;
-}
-
-.nav a:link,
-.nav a:visited {
-  text-decoration: none;
-  color: var(--color-light--2);
-  text-transform: uppercase;
-  font-size: 1.5rem;
-  font-weight: 600;
-}
-
-/* CSS Modules feature */
-.nav a:global(.active) {
-  color: var(--color-brand--2);
-}
-
-a.ctaLink:link,
-a.ctaLink:visited {
-  background-color: var(--color-brand--2);
-  color: var(--color-dark--0);
-  padding: 0.8rem 2rem;
-  border-radius: 7px;
-}
